requirements-dev.txt
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# YAML Support
pyyaml>=6.0

# Production Dependencies (aus requirements.txt)
-r huawei-solar-modbus-mqtt/requirements.txt
